Hopefully, final commit
Formatted everything, added comments, added technical documentation,
deleted the FinalProject folder from a previous version,

+++ Library/EditCreators.aspx
@@ -1,19 +1,15 @@
-﻿<%@ Page Title="" Language="VB" MasterPageFile="~/Site.master" AutoEventWireup="false" CodeFile="EditCreators.aspx.vb" Inherits="Library_EditCreators" %>
+﻿<%@ Page Title="" Language="VB" MasterPageFile="~/Site.master" AutoEventWireup="false"
+    CodeFile="EditCreators.aspx.vb" Inherits="Library_EditCreators" %>
 
-<asp:Content ID="Content1" ContentPlaceHolderID="HeadContent" Runat="Server">
+<asp:Content ID="Content1" ContentPlaceHolderID="HeadContent" runat="Server">
 </asp:Content>
-<asp:Content ID="Content2" ContentPlaceHolderID="MainContent" Runat="Server">
-    
-    <asp:Label ID="TitleLabel" runat="server" Font-Size="X-Large" 
-        Text="Edit Creators" Font-Bold="True"></asp:Label>
-    
+<asp:Content ID="Content2" ContentPlaceHolderID="MainContent" runat="Server">
+    <asp:Label ID="TitleLabel" runat="server" Font-Size="X-Large" Text="Edit Creators"
+        Font-Bold="True"></asp:Label>
     <br />
-    <asp:SqlDataSource ID="SqlDataSource" runat="server" 
-        ConnectionString="<%$ ConnectionStrings:ApplicationServices %>" 
-        DeleteCommand="DELETE FROM [Creators] WHERE [CreatorID] = @CreatorID" 
-        InsertCommand="INSERT INTO [Creators] ([Name], [Biography]) VALUES (@Name, @Biography)" 
-        SelectCommand="SELECT * FROM [Creators]" 
-        UpdateCommand="UPDATE [Creators] SET [Name] = @Name, [Biography] = @Biography WHERE [CreatorID] = @CreatorID">
+    <asp:SqlDataSource ID="SqlDataSource" runat="server" ConnectionString="<%$ ConnectionStrings:ApplicationServices %>"
+        DeleteCommand="DELETE FROM [Creators] WHERE [CreatorID] = @CreatorID" InsertCommand="INSERT INTO [Creators] ([Name], [Biography]) VALUES (@Name, @Biography)"
+        SelectCommand="SELECT * FROM [Creators]" UpdateCommand="UPDATE [Creators] SET [Name] = @Name, [Biography] = @Biography WHERE [CreatorID] = @CreatorID">
         <DeleteParameters>
             <asp:Parameter Name="CreatorID" Type="Int32" />
         </DeleteParameters>
@@ -27,13 +23,12 @@
             <asp:Parameter Name="CreatorID" Type="Int32" />
         </UpdateParameters>
     </asp:SqlDataSource>
-    <asp:ListView ID="ListView1" runat="server" DataKeyNames="CreatorID" 
-        DataSourceID="SqlDataSource" InsertItemPosition="LastItem">
+    <asp:ListView ID="ListView1" runat="server" DataKeyNames="CreatorID" DataSourceID="SqlDataSource"
+        InsertItemPosition="LastItem">
         <AlternatingItemTemplate>
-            <tr style="background-color:#FFF8DC;">
+            <tr style="background-color: #FFF8DC;">
                 <td>
-                    <asp:Button ID="DeleteButton" runat="server" CommandName="Delete" 
-                        Text="Delete" />
+                    <asp:Button ID="DeleteButton" runat="server" CommandName="Delete" Text="Delete" />
                     <asp:Button ID="EditButton" runat="server" CommandName="Edit" Text="Edit" />
                 </td>
                 <td>
@@ -48,59 +43,55 @@
             </tr>
         </AlternatingItemTemplate>
         <EditItemTemplate>
-            <tr style="background-color:#008A8C;color: #FFFFFF;">
+            <tr style="background-color: #008A8C; color: #FFFFFF;">
                 <td>
-                    <asp:Button ID="UpdateButton" runat="server" CommandName="Update" 
-                        Text="Update" />
-                    <asp:Button ID="CancelButton" runat="server" CommandName="Cancel" 
-                        Text="Cancel" />
+                    <asp:Button ID="UpdateButton" runat="server" CommandName="Update" Text="Update" />
+                    <asp:Button ID="CancelButton" runat="server" CommandName="Cancel" Text="Cancel" />
                 </td>
                 <td>
-                    <asp:Label ID="CreatorIDLabel1" runat="server" 
-                        Text='<%# Eval("CreatorID") %>' />
+                    <asp:Label ID="CreatorIDLabel1" runat="server" Text='<%# Eval("CreatorID") %>' />
                 </td>
                 <td>
                     <asp:TextBox ID="NameTextBox" runat="server" Text='<%# Bind("Name") %>' />
                 </td>
                 <td>
-                    <asp:TextBox ID="BiographyTextBox" runat="server" 
-                        Text='<%# Bind("Biography") %>' TextMode="MultiLine" />
+                    <asp:TextBox ID="BiographyTextBox" runat="server" Text='<%# Bind("Biography") %>'
+                        TextMode="MultiLine" />
                 </td>
             </tr>
         </EditItemTemplate>
         <EmptyDataTemplate>
-            <table runat="server" 
-                style="background-color: #FFFFFF;border-collapse: collapse;border-color: #999999;border-style:none;border-width:1px;">
+            <table runat="server" style="background-color: #FFFFFF; border-collapse: collapse;
+                border-color: #999999; border-style: none; border-width: 1px;">
                 <tr>
                     <td>
-                        No data was returned.</td>
+                        No data was returned.
+                    </td>
                 </tr>
             </table>
         </EmptyDataTemplate>
         <InsertItemTemplate>
             <tr style="">
                 <td>
-                    <asp:Button ID="InsertButton" runat="server" CommandName="Insert" 
-                        Text="Insert" />
-                    <asp:Button ID="CancelButton" runat="server" CommandName="Cancel" 
-                        Text="Clear" />
+                    <asp:Button ID="InsertButton" runat="server" CommandName="Insert" Text="Insert" />
+                    <asp:Button ID="CancelButton" runat="server" CommandName="Cancel" Text="Clear" />
                 </td>
                 <td>
-                    &nbsp;</td>
+                    &nbsp;
+                </td>
                 <td>
                     <asp:TextBox ID="NameTextBox" runat="server" Text='<%# Bind("Name") %>' />
                 </td>
                 <td>
-                    <asp:TextBox ID="BiographyTextBox" runat="server" 
-                        Text='<%# Bind("Biography") %>' TextMode="MultiLine" />
+                    <asp:TextBox ID="BiographyTextBox" runat="server" Text='<%# Bind("Biography") %>'
+                        TextMode="MultiLine" />
                 </td>
             </tr>
         </InsertItemTemplate>
         <ItemTemplate>
-            <tr style="background-color:#DCDCDC;color: #000000;">
+            <tr style="background-color: #DCDCDC; color: #000000;">
                 <td>
-                    <asp:Button ID="DeleteButton" runat="server" CommandName="Delete" 
-                        Text="Delete" />
+                    <asp:Button ID="DeleteButton" runat="server" CommandName="Delete" Text="Delete" />
                     <asp:Button ID="EditButton" runat="server" CommandName="Edit" Text="Edit" />
                 </td>
                 <td>
@@ -118,33 +109,37 @@
             <table runat="server">
                 <tr runat="server">
                     <td runat="server">
-                        <table ID="itemPlaceholderContainer" runat="server" border="1" 
-                            style="background-color: #FFFFFF;border-collapse: collapse;border-color: #999999;border-style:none;border-width:1px;font-family: Verdana, Arial, Helvetica, sans-serif;">
-                            <tr runat="server" style="background-color:#DCDCDC;color: #000000;">
+                        <table id="itemPlaceholderContainer" runat="server" border="1" style="background-color: #FFFFFF;
+                            border-collapse: collapse; border-color: #999999; border-style: none; border-width: 1px;
+                            font-family: Verdana, Arial, Helvetica, sans-serif;">
+                            <tr runat="server" style="background-color: #DCDCDC; color: #000000;">
                                 <th runat="server">
                                 </th>
                                 <th runat="server">
-                                    CreatorID</th>
+                                    CreatorID
+                                </th>
                                 <th runat="server">
-                                    Name</th>
+                                    Name
+                                </th>
                                 <th runat="server">
-                                    Biography</th>
+                                    Biography
+                                </th>
                             </tr>
-                            <tr ID="itemPlaceholder" runat="server">
+                            <tr id="itemPlaceholder" runat="server">
                             </tr>
                         </table>
                     </td>
                 </tr>
                 <tr runat="server">
-                    <td runat="server" 
-                        style="text-align: center;background-color: #CCCCCC;font-family: Verdana, Arial, Helvetica, sans-serif;color: #000000;">
+                    <td runat="server" style="text-align: center; background-color: #CCCCCC; font-family: Verdana, Arial, Helvetica, sans-serif;
+                        color: #000000;">
                         <asp:DataPager ID="DataPager1" runat="server">
                             <Fields>
-                                <asp:NextPreviousPagerField ButtonType="Button" ShowFirstPageButton="True" 
-                                    ShowNextPageButton="False" ShowPreviousPageButton="False" />
+                                <asp:NextPreviousPagerField ButtonType="Button" ShowFirstPageButton="True" ShowNextPageButton="False"
+                                    ShowPreviousPageButton="False" />
                                 <asp:NumericPagerField />
-                                <asp:NextPreviousPagerField ButtonType="Button" ShowLastPageButton="True" 
-                                    ShowNextPageButton="False" ShowPreviousPageButton="False" />
+                                <asp:NextPreviousPagerField ButtonType="Button" ShowLastPageButton="True" ShowNextPageButton="False"
+                                    ShowPreviousPageButton="False" />
                             </Fields>
                         </asp:DataPager>
                     </td>
@@ -152,10 +147,9 @@
             </table>
         </LayoutTemplate>
         <SelectedItemTemplate>
-            <tr style="background-color:#008A8C;font-weight: bold;color: #FFFFFF;">
+            <tr style="background-color: #008A8C; font-weight: bold; color: #FFFFFF;">
                 <td>
-                    <asp:Button ID="DeleteButton" runat="server" CommandName="Delete" 
-                        Text="Delete" />
+                    <asp:Button ID="DeleteButton" runat="server" CommandName="Delete" Text="Delete" />
                     <asp:Button ID="EditButton" runat="server" CommandName="Edit" Text="Edit" />
                 </td>
                 <td>
@@ -171,4 +165,3 @@
         </SelectedItemTemplate>
     </asp:ListView>
 </asp:Content>
-
